// packages/solid/src/additional-components/NodeToolbar/types.ts
import type { JSX } from 'solid-js';
import type { Position, Align } from '@xyflow/system';

/**
 * @expand
 */
export type NodeToolbarProps = Omit<JSX.HTMLAttributes<HTMLDivElement>, 'children'> & {
  /** Id of the node, or array of ids the toolbar should be displayed at */
  nodeId?: string | string[];
  /** If true, node toolbar is visible even if node is not selected */
  isVisible?: boolean;
  /**
   * Position of the toolbar relative to the node
   * @example Position.TopLeft, Position.TopRight,
   * Position.BottomLeft, Position.BottomRight
   */
  position?: Position;
  /** Offset the toolbar from the node */
  offset?: number;
  /**
   * Align the toolbar relative to the node
   * @example Align.Start, Align.Center, Align.End
   */
  align?: Align;
  /** The content to render in the toolbar */
  children?: JSX.Element;
};